+++ lib/context_extensions_codespark.dart
@@ -1,5 +1,8 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/extensions/media_query_extensions.dart';
+export 'src/extensions/device_extensions.dart';
+export 'src/extensions/theme_extensions.dart';
+export 'src/extensions/typography_extensions.dart';
+export 'src/extensions/snackbar_extensions.dart';
+export 'src/extensions/focus_extensions.dart';
+export 'src/extensions/lifecycle_extensions.dart';
+export 'src/extensions/orientation_extensions.dart';